Improved UI and interaction with console
Moved all console management to a Screen Class which also handles all output to the screen. It will contain a 2D vector of pointers to selectable items. The vector components correspond to each console cell on the screen. This will allow clicking on the screen to select a Selectable object.