More for GameInstance and some tidying
Made it so that there is a single instance of the GameInstance class that is globally accessible.
The class has member functions to return the current screen and map. These are therefore now globally accessible through the GameInstance. Meaning that I don't need to initiate classes and functions with the display and map so they can get access to the gamestate, they can now just always do that anyway.
I also tidied up the headers and forward declarations and went through all functions and made sure they had appropriate consting and passed by reference where possible.