Movement animation
Added animation to unit movement. One function which does single tile movement and another function which takes a MoveSequence object and performs the movement + animation.
Also made a function in map  which returns a tile when given a console coordinate. This tidied the code up a bit and reduced repitition