Implemented context/action menu
Implemented a context menu that is dictates the options available to the player and structures the flow of the game.
When user selects a unit they must pick an action from the options menu which they can make after pressing a text button on the menu.
Updated unit representation so that units have 2 chars to identify them and their hp is given on their tile too.