Smart pointers, rule of 5, and more const
Replaced some pointers with unique pointers to establish clearer ownership of pointers and make cleanup easier.
Checked which classes needed any of "rule of 5" functions and created the remaining functions if necessary (Map, GameInstance).
Went through all functions and determined if constness could be brought in to make code safer (for input, output, and this).