Some code reorganising
Moved the tile_width/height and map_x/y_offset variables from the Window class to the Map class as makes more sense to have there.
GameInstance no longer has a pointer to the Window as a member. Instead access to the window is obtained through the static GameManager object